refactor: move supporting types into module
